Add Mozjpeg 3.0 speed optimizations.
This change fixes a regression introduced by the upgrade to Mozjpeg 3.0 by
setting some extra values in the compression structure when `optimize` mode
is off, which is done during the binary search for the best visual quality.
Speed is improved about 2x in local tests, and compared to the latest
official jpeg-archive release the speed is slightly faster and the produced
output has a slightly smaller file size at the same visual quality.